Correction de la doc de KOfamScan
Predited KEGG Orthologs -> KOfamScan
The KO assignment of a gene allow to identify a function and define a molecular networks. ->
The KO assignment of a gene allows to identify a function and define a molecular network.
we use KofamScan -> we use KOfamScan
In MicroScope, only the results that matches with a score exceeding the pre-computed threshold and a E-value lower than or equal to 0.01 are visualizable. ->
In MicroScope, only the results that match with a score exceeding the pre-computed threshold and a E-value lower than or equal to 0.01 are shown.
KO : ID number of KEGG Orthologs entry -> KO : ID number of the KEGG Ortholog entry
Eval : E value of the match -> Eval : E-value of the match
Functional description of KO -> Functional description of the KO
EC number : EC number associated with KO -> EC number : EC number associated with the KO